backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary
Pillow>=9.5.0
pytest==8.3.3
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-xdist==3.6.1
pytest-env==1.1.5
pytest-mock==3.14.0
PyYAML>=6.0
